updated user model and login/reg components
