+++ LifeSignsFront/src/app/register/register.component.html
@@ -7,33 +7,55 @@
               <div class="card-body p-5">
                 <h2 class="text-uppercase text-center mb-5">Create an account</h2>
   
-                <form>
-  
+                <form [formGroup]="registerForm">
+
                   <div class="form-outline mb-4">
-                    <input type="text" id="form3Example1cg" class="form-control form-control-lg" />
-                    <label class="form-label" for="form3Example1cg">User Name</label>
+                    <label class="form-label" for="form3Example3cg"><b>Account Type: </b></label>
+                    <label id="asterisk">* </label> 
+                    <select formControlName="role">
+                      <option value="" disabled selected>Select an option</option>
+                      <option value="Doctor">Doctor</option>
+                      <option value="Nurse">Nurse</option>
+                      <option value="Patient">Patient</option>
+                    </select>
+                  </div>
+
+                  <div class="form-outline mb-4">
+                    <label class="form-label" for="form3Example3cg"><b>Email: </b></label>
+                    <label id="asterisk">* </label> 
+                    <input formControlName="email" type="email" id="form3Example3cg" class="form-control form-control-lg" />
                   </div>
   
                   <div class="form-outline mb-4">
-                    <input type="email" id="form3Example3cg" class="form-control form-control-lg" />
-                    <label class="form-label" for="form3Example3cg">Your Email</label>
+                    <label class="form-label" for="form3Example3cg"><b>Username: </b></label>
+                    <label id="asterisk">* </label> 
+                    <input formControlName="username" type="text" id="form3Example1cg" class="form-control form-control-lg" />
                   </div>
   
                   <div class="form-outline mb-4">
-                    <input type="password" id="form3Example4cg" class="form-control form-control-lg" />
-                    <label class="form-label" for="form3Example4cg">Password</label>
+                    <label class="form-label" for="form3Example3cg"><b>Password: </b></label>
+                    <label id="asterisk">* </label> 
+                    <input formControlName="password" type="password" id="form3Example4cg" class="form-control form-control-lg" />
                   </div>
   
                   <div class="form-outline mb-4">
-                    <input type="password" id="form3Example4cdg" class="form-control form-control-lg" />
-                    <label class="form-label" for="form3Example4cdg">Repeat your password</label>
+                    <label class="form-label" for="form3Example3cg"><b>Confirm Password: </b></label>
+                    <label id="asterisk">* </label> 
+                    <input formControlName="confirmed" type="password" id="form3Example4cdg" class="form-control form-control-lg" />
+                  </div>
+
+                  <div class="form-outline mb-4">
+                    <label id="asterisk">* </label> 
+                    <label id="required-text"><small>Required</small></label>
                   </div>
   
                   <div class="d-flex justify-content-center">
-                    <button type="button" class="btn btn-success btn-block btn-lg gradient-custom-4 text-body">Register</button>
+                    <button type="submit" (click)="submitUser(registerForm)" class="btn btn-success btn-block btn-lg gradient-custom-4 text-body">
+                      Register
+                    </button>
                   </div>
   
-                  <p class="text-center text-muted mt-5 mb-0">Have already an account? <a href="#!" class="fw-bold text-body"><u>Login here</u></a></p>
+                  <p class="text-center text-muted mt-5 mb-0">Already have an account? <a href="#!" class="fw-bold text-body"><u>Login here</u></a></p>
   
                 </form>
